LineageOS-18.1-Ubuntu-Kernel-SM8250
LineageOS-18.1-Ubuntu-Kernel-SM8250

+++ drivers/platform/msm/ipa/ipa_v3/ipahal/ipahal_hw_stats.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-2.0-only
 /*
- * Copyright (c) 2017-2019, 2021, The Linux Foundation. All rights reserved.
+ * Copyright (c) 2017-2019, The Linux Foundation. All rights reserved.
  */
 
 #include "ipahal.h"
@@ -225,11 +225,11 @@ static struct ipahal_stats_init_pyld *ipahal_generate_init_pyld_flt_rt_v4_5(
 	void *params, bool is_atomic_ctx)
 {
 	struct ipahal_stats_init_pyld *pyld;
-	long num = (long)(params);
+	int num = (int)(params);
 
 	if (num > IPA_MAX_FLT_RT_CNT_INDEX ||
 		num <= 0) {
-		IPAHAL_ERR("num %ld not valid\n", num);
+		IPAHAL_ERR("num %d not valid\n", num);
 		return NULL;
 	}
 	pyld = IPAHAL_MEM_ALLOC(sizeof(*pyld) +
